[CAL-04] refactor: Viewer 계층 수정
- 상세 로직들을 메서드로 추출
- 주석 추가
