--- .idea/workspace.xml
@@ -1,115 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="AutoImportSettings">
-    <option name="autoReloadType" value="SELECTIVE" />
-  </component>
-  <component name="ChangeListManager">
-    <list default="true" id="848ddfe2-bc99-4689-99f9-8d94ca1ccc66" name="Changes" comment="[CAL-02] feat: Viewer 계층 구현&#10;&#10;- 구체 클래스는 Viewer 인터페이스의 run() 메서드를 구현하도록 설계&#10;- run() 메서드에서 사용자 입출력에 대한 로직 구현&#10;- Console 환경을 기반으로 입출력을 처리하는 ConsoleViewer() 구현&#10;- 사용자 입력 상태를 Enum으로 관리" />
-    <option name="SHOW_DIALOG" value="false" />
-    <option name="HIGHLIGHT_CONFLICTS" value="true" />
-    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
-    <option name="LAST_RESOLUTION" value="IGNORE" />
-  </component>
-  <component name="ExternalProjectsData">
-    <projectState path="$PROJECT_DIR$">
-      <ProjectState />
-    </projectState>
-  </component>
-  <component name="FileTemplateManagerImpl">
-    <option name="RECENT_TEMPLATES">
-      <list>
-        <option value="Enum" />
-      </list>
-    </option>
-  </component>
-  <component name="Git.Settings">
-    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
-  </component>
-  <component name="MarkdownSettingsMigration">
-    <option name="stateVersion" value="1" />
-  </component>
-  <component name="ProjectId" id="2GZMd7Jw5MQqqswgL0KSeA2b5SW" />
-  <component name="ProjectLevelVcsManager" settingsEditedManually="true">
-    <ConfirmationsSetting value="1" id="Add" />
-  </component>
-  <component name="ProjectViewState">
-    <option name="showLibraryContents" value="true" />
-  </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "WebServerToolWindowFactoryState": "false",
-    "last_opened_file_path": "C:/Users/HwanKim/projects/java-calculator/build.gradle",
-    "nodejs_package_manager_path": "npm",
-    "settings.editor.selected.configurable": "reference.settingsdialog.project.gradle"
-  }
-}]]></component>
-  <component name="RunManager">
-    <configuration name="Main" type="Application" factoryName="Application" temporary="true" nameIsGenerated="true">
-      <option name="MAIN_CLASS_NAME" value="calculator.Main" />
-      <module name="java-calculator.main" />
-      <extension name="coverage">
-        <pattern>
-          <option name="PATTERN" value="calculator.*" />
-          <option name="ENABLED" value="true" />
-        </pattern>
-      </extension>
-      <method v="2">
-        <option name="Make" enabled="true" />
-      </method>
-    </configuration>
-    <recent_temporary>
-      <list>
-        <item itemvalue="Application.Main" />
-      </list>
-    </recent_temporary>
-  </component>
-  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
-  <component name="TaskManager">
-    <task active="true" id="Default" summary="Default task">
-      <changelist id="848ddfe2-bc99-4689-99f9-8d94ca1ccc66" name="Changes" comment="" />
-      <created>1666591750256</created>
-      <option name="number" value="Default" />
-      <option name="presentableId" value="Default" />
-      <updated>1666591750256</updated>
-      <workItem from="1666591755074" duration="39000" />
-      <workItem from="1666592775335" duration="3441000" />
-    </task>
-    <task id="LOCAL-00001" summary="[CAL-02] chore: 폴더 구조 변경">
-      <created>1666592886481</created>
-      <option name="number" value="00001" />
-      <option name="presentableId" value="LOCAL-00001" />
-      <option name="project" value="LOCAL" />
-      <updated>1666592886481</updated>
-    </task>
-    <task id="LOCAL-00002" summary="[CAL-02] feat: Viewer 계층 구현&#10;&#10;- 구체 클래스는 Viewer 인터페이스의 run() 메서드를 구현하도록 설계&#10;- run() 메서드에서 사용자 입출력에 대한 로직 구현&#10;- Console 환경을 기반으로 입출력을 처리하는 ConsoleViewer() 구현&#10;- 사용자 입력 상태를 Enum으로 관리">
-      <created>1666596085331</created>
-      <option name="number" value="00002" />
-      <option name="presentableId" value="LOCAL-00002" />
-      <option name="project" value="LOCAL" />
-      <updated>1666596085331</updated>
-    </task>
-    <option name="localTasksCounter" value="3" />
-    <servers />
-  </component>
-  <component name="TypeScriptGeneratedFilesManager">
-    <option name="version" value="3" />
-  </component>
-  <component name="Vcs.Log.Tabs.Properties">
-    <option name="TAB_STATES">
-      <map>
-        <entry key="MAIN">
-          <value>
-            <State />
-          </value>
-        </entry>
-      </map>
-    </option>
-  </component>
-  <component name="VcsManagerConfiguration">
-    <MESSAGE value="[CAL-02] chore: 폴더 구조 변경" />
-    <MESSAGE value="[CAL-02] feat: Viewer 계층 구현&#10;&#10;- 구체 클래스는 Viewer 인터페이스의 run() 메서드를 구현하도록 설계&#10;- run() 메서드에서 사용자 입출력에 대한 로직 구현&#10;- Console 환경을 기반으로 입출력을 처리하는 ConsoleViewer() 구현&#10;- 사용자 입력 상태를 Enum으로 관리" />
-    <option name="LAST_COMMIT_MESSAGE" value="[CAL-02] feat: Viewer 계층 구현&#10;&#10;- 구체 클래스는 Viewer 인터페이스의 run() 메서드를 구현하도록 설계&#10;- run() 메서드에서 사용자 입출력에 대한 로직 구현&#10;- Console 환경을 기반으로 입출력을 처리하는 ConsoleViewer() 구현&#10;- 사용자 입력 상태를 Enum으로 관리" />
-  </component>
-</project>
