+++ src/main/java/calculator/console/Viewer.java
@@ -1,6 +1,6 @@
 package calculator.console;
 
-// 입출력을 담당하는 interface
+// 사용자 UI(입출력)을 담당하는 interface
 public interface Viewer {
 
     // 계산기 console을 실행하는 뼈대가 되는 메서드입니다.
